+++ packages/graphql_codegen/test/assets/graphql_flutter_client_query_no_arguments/document.graphql.dart
@@ -149,9 +149,7 @@ class Options$Query$FetchSNoVariables
               ? null
               : (data) => onComplete(
                     data,
-                    data == null
-                        ? null
-                        : _parserFn$Query$FetchSNoVariables(data),
+                    _parserFn$Query$FetchSNoVariables(data),
                   ),
           onError: onError,
           document: documentNodeQueryFetchSNoVariables,