+++ packages/graphql_codegen/test/assets/graphql_flutter_client_query_optional_arguments/document.graphql.dart
@@ -259,7 +259,7 @@ class Options$Query$FetchSOptional
               ? null
               : (data) => onComplete(
                     data,
-                    data == null ? null : _parserFn$Query$FetchSOptional(data),
+                    _parserFn$Query$FetchSOptional(data),
                   ),
           onError: onError,
           document: documentNodeQueryFetchSOptional,