+++ packages/graphql_codegen/test/assets/graphql_flutter_client_query_required_arguments/document.graphql.dart
@@ -252,7 +252,7 @@ class Options$Query$FetchSRequired
               ? null
               : (data) => onComplete(
                     data,
-                    data == null ? null : _parserFn$Query$FetchSRequired(data),
+                    _parserFn$Query$FetchSRequired(data),
                   ),
           onError: onError,
           document: documentNodeQueryFetchSRequired,